fix: service

+++ run_agent.sh
@@ -1,4 +1,8 @@
-rm -r celo_trader
+if test -d celo_trader; then
+  echo "Removing previous agent build"
+  rm -r celo_trader
+fi
+
 find . -empty -type d -delete  # remove empty directories to avoid wrong hashes
 autonomy packages lock
 autonomy fetch --local --agent valory/celo_trader && cd celo_trader
